added subtitle

+++ workflow/src/utils/supabase_utils.py
@@ -10,11 +10,18 @@ def checkIfExists(column: str, value: any) -> bool:
     result = searchRow(SUPABASE_TABLE, column, value)
     return len(result) > 0
 
-def searchRow(tableName: str, searchColumn: str, searchValue: any):
+def searchRow(tableName: str, searchColumn: str, searchValue: any, filter_column: str = None, filter_value: any = None):
     """Search for a row in the database."""
     try:
-        data = supabase.table(tableName).select("*").eq(searchColumn, searchValue).execute()
-        # print(f"searchRow result: {data.data}")
+        query = supabase.table(tableName).select("*").eq(searchColumn, searchValue).order("created_at", desc=True)
+        
+        # If filtering for NULL values
+        if filter_column and filter_value is None:
+            query = query.is_(filter_column, 'null')
+        # For non-NULL filter values
+        elif filter_column and filter_value is not None:
+            query = query.eq(filter_column, filter_value)
+        data = query.execute()
         return data.data
     except Exception as error:
         print(f"Error searching for a row: {error}")
@@ -53,28 +60,22 @@ def updateRow(updates: dict, searchColumn: str, searchValue: int):
 def upload_audio_file(file_path: str, bucket_name: str = SUPABASE_BUCKET_NAME) -> str:
     """
     Upload an audio file to Supabase storage and return the public URL.
-    
-    Args:
-        file_path: Path to the audio file
-        bucket_name: Name of the storage bucket (default: SUPABASE_BUCKET_NAME)
-    
-    Returns:
-        str: Public URL of the uploaded file
     """
     logger = get_run_logger()
     try:
         # Get the filename from the path
         file_name = Path(file_path).name
+        print(f"upload_audio_file file_name: {file_path}")
         # Read the file
         with open(file_path, 'rb') as f:
             file_data = f.read()
         # Upload to Supabase storage
-        response = supabase.storage.from_(bucket_name).upload(
+        supabase.storage.from_(bucket_name).upload(
             path=file_name,
             file=file_data,
             file_options={"content-type": "audio/mpeg"}
         )
-        print(f"upload_audio_file response: {response}")
+        
         # Get the public URL
         public_url = supabase.storage.from_(bucket_name).get_public_url(file_name)
         print(f"File uploaded successfully: {public_url}")
@@ -84,4 +85,5 @@ def upload_audio_file(file_path: str, bucket_name: str = SUPABASE_BUCKET_NAME) -
         logger.error(f"Error uploading audio file: {error}")
     finally:
         # delete the file from local storage
-        Path(file_path).unlink()
+        if Path(file_path).exists():
+            Path(file_path).unlink()
